feat: add swagger and its config and also modify user and token model to have role

+++ src/main/java/com/example/fintech/model/token/dao/TokenRepository.java
@@ -11,9 +11,9 @@ import org.springframework.data.jpa.repository.Query;
 public interface TokenRepository extends JpaRepository<Token, Integer> {
 
 	@Query("""
-			select t from Token  t inner join User u on t.user.id=u.id where u.id= :userId and (t.expired= false or t.revoked=false ) 
+			select t from Token t inner join User u on t.user.id=u.id where u.id= :userId and (t.expired= false or t.revoked=false ) 
 			""")
-	List<Token> findAllValidTokensByUser(Integer userId);
+	List<Token> findAllValidTokensByUser(Long userId);
 
 	Optional<Token> findByToken(String token);
